+++ src/app/components/SettingsMenu/index.tsx
@@ -115,7 +115,7 @@ const listStyle: React.CSSProperties = {
     overflowY: 'visible',
 };
 
-export const C_VERSION = '0.30.1';
+export const C_VERSION = '0.30.3';
 export const C_CUSTOM_BG_ID = 'river_custom_bg';
 export const C_AVATAR_SIZE = 640;